Included a check when adding string to description
When we have comments that span multiple lines, we display it inline and
thus we need a space between the lines, otherwise last word of line X will be
glued to first word of line X+1.

Instead of just adding space always when appending description, we need
to check if there's something there already, otherwise it's just an
empty string and our description would start with a blank space.

Solves: https://github.com/sroze/companienv/issues/17